Move the `LinkedInputs` functionality into the `ExperimentRunConfig` class
