+++ ert3/config/__init__.py
@@ -11,8 +11,7 @@ from ._stages_config import (
 from ._validator import DEFAULT_RECORD_MIME_TYPE
 from ._experiment_config import load_experiment_config, ExperimentConfig
 from ._parameters_config import load_parameters_config, ParametersConfig
-from ._linked_inputs import LinkedInput, link_inputs
-from ._experiment_run_config import ExperimentRunConfig
+from ._experiment_run_config import ExperimentRunConfig, LinkedInput
 
 __all__ = [
     "load_ensemble_config",
@@ -31,6 +30,5 @@ __all__ = [
     "IndexedOrderedDict",
     "TransportableCommand",
     "LinkedInput",
-    "link_inputs",
     "ExperimentRunConfig",
 ]
